Archive: DEPARTMENT OF COMPUTER SCIENCE
Archive of the official website of DEPARTMENT OF COMPUTER SCIENCE of Kabul university

Now the Computer Science department doesn't have student. Instructors of the Computer Science department are teaching computer science subjects in the faculty first year students and other faculties of Kabul University. Currently they are working to develop a curriculum in order to graduate students.